[packages] aiccu: remove complete integration, only leave the binary
According to its own author, aiccu is broken on nat gateways, not meant for
unattended operation and in the need for a hardware rtc to function properly.

Since restarting aiccu or attempting to work around those problems violate
the sixxs terms of service, this commit removes any non-necessary patches
and any scripts shipped with it.

Users having trouble with aiccu may contact the sixxs staff for assistance.

git-svn-id: svn://svn.openwrt.org/openwrt/packages@32666 3c298f89-4303-0410-b956-a3cf2f4a3e73