[packages] privoxy: move privoxy startup to later in boot process
When using privoxy as a transparent HTTP proxy a race condition occurs
where privoxy may attempt to startup and start listening on the LAN
interface before the network configuration is completed. As a result
privoxy fails to start up because it can't bind to the correct listening
address. This patch changes the init.d script to be a little bit later
in the boot process.

Signed-off-by: Adam Gensler <openwrt@kristenandadam.net>

git-svn-id: svn://svn.openwrt.org/openwrt/packages@32759 3c298f89-4303-0410-b956-a3cf2f4a3e73