[package] add ostiary package
This patch adds the Ostiary package to the repository.  I've tested this on the latest build (revision 31158).  Installation, un-installation and ostiaryd.init script all seem to work correctly.

A bit of background, Ostiary has a similar purpose to Port Knocking but uses a different implementation that removes the possibility of replay attacks.  Not everyone will view it as suitable for their needs, but for me combined with ssh I feel it is more than sufficient.

I suppose for the super paranoid admin, the perfect system would be a package that uses the approach of port knocking via a closed port that is combined with an ostiary style signature challenge.

Anyhow, I have created quickstart documentation for the server and client at "http://wiki.openwrt.org/doc/howto/ostiary.server" and  "http://wiki.openwrt.org/doc/howto/ostiary.client".

Signed-off-by: Desmond Hokin <Salander@gmx.u>

git-svn-id: svn://svn.openwrt.org/openwrt/packages@33167 3c298f89-4303-0410-b956-a3cf2f4a3e73