[bluez 4.99]  Added a new feed for bluez 4.99
Bluez 4.99 moves to a new model which combines -lib and -utils.
For the time being there will be all 3 in the package feeds list,
but we need 4.99 to add in A2DP support so we need it.  I do not
recommend selecting bluez bluez-libs and bluez-utils simultaneously
the world my explode.  We'll remove the older feeds once we know the
new feed works.

Change-Id: If710cfaecc7d4513c597e47b97c475a5d68509f9
Signed-off-by: Michael Larson <mlarson@codeaurora.org>
Reviewed-by: Mathieu Olivari <mathieu@codeaurora.org>