[xtables-addons] Fix build issue on msm kernel
Building with msm-3.4 kernel creates a build issue
"SHRT_MAX already defined in linux/kernel.h". Though the reason
why it works on the openwrt 3.3 kernel is unclear, it's fixed
by moving the headers and making sure the kernel header is
included before redefining this function. xtables-addons makes
sure the macro is not already defined before defining it
locally.

Change-Id: If915c49dbb7e9fa0bba878175621fecb84273e6a
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>