[libgpg-error][lzo] add libtool depedency to build
This has the effect of making libtool a PKG_BUILD_DEPENDS, so it will be
done being built by the time this package starts building. The problem
scenario is the following:

 make[3] -C package/libtool compile
 make[3] -C qca/feeds/packages/libs/libgpg-error compile
 make -C package/libtool compile BUILD_VARIANT= finished
libtool: Version mismatch error.  This is libtool 2.4, but the
libtool: definition of this LT_INIT comes from libtool 2.2.6.
libtool: You should recreate aclocal.m4 with macros from libtool 2.4
libtool: and run autoconf again.
make[7]: *** [libgpg_error_la-init.lo] Error 63
 make -C package/feeds/packages/libgpg-error compile BUILD_VARIANT=
failed

It's quite hard to verify this will fix things. We will keep watching
for build issues with this pattern

Change-Id: If76f8f6d4ea731f7fa4d130bb10cf7eebe23682a
Signed-off-by: Matthew McClintock <mmcclint@codeaurora.org>
Reviewed-by: Mathieu Olivari <mathieu@codeaurora.org>