[hiredis] don't put extra "" in var
The makefile for this package is done poorly to begin with wrt
to using this variable, but it also seems some version of bash
will interpret this being set to ARCH="" as an extra argument.

This really throws things off when invoking the compiler as it's
trying to compile a file name "".

Change-Id: I049ff2e799928d778834beac65c19f24fcd13553
Signed-off-by: Matthew McClintock <mmcclint@codeaurora.org>