[miniupnpd] fixes bugs of CDRouter
In the previous picks, it use a non-exist command "fw3"in the current
depot, the fw3 command is the new implementation of firewall which is
not merged yet.
and fixes some CDRouter bugs caused by the restart of the daemon,
to fixes this bug, there will not restart the daemon when it already
is running.

Change-Id: Ie13b3e91a32e62688ae37f9a9da708d3b730401f
Signed-off-by: Zhu Ken <guigenz@codeaurora.org>
Reviewed-by: Mathieu Olivari <mathieu@codeaurora.org>