[perf] Add External tree build support
When building with an external tree, QSDK uses the O= option, which
results in having the binaries outside of the source tree. A new
variable $(LINUX_SRC_DIR) have been created to specifically point
to the source.

This change replaces LINUX_DIR in that package with this new variable
so perf can compile on linux-3.4 kernel.

Change-Id: Ib3edcf3e64c864840eabd2a609184ba0dec0f6b0
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>