[miniupnpd] handle overflow when read 64bit counter
CR's Fixed: 557774, 557775
counters in /proc/net/dev is 64bit counter, miniupnpd use function
"strtoul" to convert the output string to a 32bit integer,but function
"strtoul" always return 4294967295 when the 64bit counter is larger
than 4294967295. so i replace "strtoul" with "strtoull" to handle
overflow when reading a 64bit counter string.

Change-Id: Ie1feacadeddae81a86e0d037f6399a3d8e5a8591
Signed-off-by: Xiaoping Fan <xfan@codeaurora.org>