[mplayer] Force libfaad disable in the config stage
By default, mplayer ./configure automatically detects the presence of
libraries and enable features accordingly. So it's a problem if a
library is in staging_dir due to a previous build or another package
dependency. In that case, mplayer will be built with the support for
that library but without explicitely listing it as a dependency, which
will trigger an error in OpenWrt.

This fixes the following error:

Package mplayer is missing dependencies for the following libraries:
libfaad.so.2

Change-Id: I2262bda9a1b67a8c3531b3f622186cf773861684
Signed-off-by: Matthew McClintock <mmcclint@codeaurora.org>
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>