[xl2tpd] force redial when pppd exit
CRs-Fixed:  652225
When pppd exit either for manually killing or keep alive packet losing,
xl2tpd will send out a call disconnect notify(CDN) to l2tp network
server(LNS), and wait for an ack(ZLB) on that call. But the session id
of the ZLB ack from LNS is 0, it maps to control connection which
needn't to close. So the call which need to close will always wait for
its closing ack while its expected ack has been sent to control
connection which needn't to close. This result in a incomplete cleanup,
and the l2tp tunnel can't recover by redial.

Here i set needclose to false and closing to true when pppd exit to
force xl2tpd to completely clean the call, and redial after 20 seconds.

Change-Id: Iafff5f2b0d78ccac47c783e6af01137b9c256f1e
Signed-off-by: Xiaoping Fan <xfan@codeaurora.org>