[mcproxy] Remove coredump files
CRs-Fixed: 689700
Mcproxy would exit with signal SIGABRT when network interfaces are not prepared.
This often occurs when networking is restarting. The root cause is GCC4.6 didn't
handle throw/catch well when there are destructors of class members or local
variables.
When mcproxy tried to exit with a warning message(by throwing an exception),
the exception was not catched. That is why mcproxy exited with SIGABRT.
This is a workaroud for the compiler issue.

Change-Id: I03a0b93e9813113134a70ea0b7574b6cbcdf6908
Signed-off-by: Zhi Chen <zhichen@codeaurora.org>