[libnl-bf] add dummy data to blackhole nlmsg
added a dummy value to the blackhole qdisc type nl msg.

the function rtnl_tc_set_kind issues a BUG() if the priv size returned
by a qdisc's to_size op is zero.  simply removing the BUG() and returning
from the function was also tried, but it made the test system unstable,
which appears to be because many places in the libnl code require
a non-zero data space for qdisc messages.

Change-Id: Icaf71347b77ccb5941082aca8b3f49004c4a767a
Signed-off-by: Nicolas Costa <ncosta@codeaurora.org>