mcproxy: fix unexpected core dump
CRs-Fixed: 838511
The try/catch cannot get some exceptions when thread was used. That
caused the application aborted with core dump. It's related to libc
and pthead implementation(not all platforms have such an issue).
This fixing used the compiling option "-pthread" instead of "-lpthread".
Besides of linking to pthread, it enabled reentrant APIs in libc, so
that the exceptions could be handled properly in MT context.

Change-Id: I2c1bb94148f4741a4b6b21a9e7e4cb9a118b64cd
Signed-off-by: Zhi Chen <zhichen@codeaurora.org>