[rp-pppoe]: the server don't use the right plugin location
In the rp-pppoe server, it will use a referred rp-pppoe.so plugin,
but this plugin is not fit to the pppd, we need use the rp-pppoe.so
plugin from pppd package, we need not rp-pppoe package's plugin.

remove the copy the rp-pppoe.so from the rp-pppoe package, let the pppd
use the default rp-pppoe plugin from pppd package.

Change-Id: I805e516e2b7a3277cbe64cb0d11a90a22a8bc817
Signed-off-by: Zhu Ken <guigenz@codeaurora.org>