[openswan-nss]when initiator stop,it don't send out the state change
when the initiator stop, it don't send out the state change event
to the responder, it cause that respondor won't change its state
correctly, when it start again, the respondor will just update
the state without notify firewall.

it only send out notify when ikv1 is present, change it to send out
notify as well for ikv2.

Change-Id: I1e3ce7f31407a06db9ff5ab2e317e097181759c2
Signed-off-by: Zhu Ken <guigenz@codeaurora.org>